Updated python and libraries
Updated to 3.14
Updated numpy and scipy; Fixed resulting errors

+++ ADCP.py
@@ -858,7 +858,7 @@ def Inverse(
     # Dv = [spdiags(repmat([-1 2 -1], Nt-2,1),[0 1 2], Nt-2,Nt), sparse(Nt-2,2*Nz) ]*VEH_SMOOTH  ; % d/dt
 
     diags = np.tile(np.array([-1, 2, -1]), (Nt - 2, 1))
-    dd_dv = sp.sparse.diags_array(diags.T, offsets=[0, 1, 2], shape=(Nt - 2, Nt))
+    dd_dv = sp.sparse.diags_array(diags.T, offsets=[0, 1, 2], shape=(Nt - 2, Nt), dtype=np.float64)
     Dv = sp.sparse.hstack([dd_dv, sp.sparse.csr_array((Nt - 2, 2 * Nz))]) * weights.VEH_SMOOTH  # d/dt
     if inverse_tmp is not None:
         inverse_tmp["Dv"] = Dv.todense()
